fixed typed dict location
